refactor: reorganize project
