+++ src/main.py
@@ -1,13 +1,16 @@
-from . import app
 import argparse
+from downloader import start_download
 
-if __name__ == "__main__":
+def main():
     parser = argparse.ArgumentParser(description="YouTube Downloader")
     parser.add_argument(
         "--audio-only",
         action="store_true",
-        help="Download only audio in MP3 format.",
+        help="Download only the audio as MP3"
     )
     args = parser.parse_args()
 
-    app.start_download(audio_only=args.audio_only)
+    start_download(audio_only=args.audio_only)
+
+if __name__ == "__main__":
+    main()
